Integrated Mongodb Atlas to the app

+++ src/main/java/com/surya/journalApp/Service/JournalEntryService.java
@@ -1,11 +1,13 @@
 package com.surya.journalApp.Service;
 
 import com.surya.journalApp.Entity.JournalEntry;
+import com.surya.journalApp.Entity.User;
 import com.surya.journalApp.Repository.JournalEntryRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -16,16 +18,28 @@ public class JournalEntryService {
 
     @Autowired
    private JournalEntryRepository journalEntryRepository;
+    @Autowired
+    private UserService userService;
 
-    public void saveEntry(JournalEntry journalEntry) {
+
+    @Transactional
+    public void saveEntry(JournalEntry journalEntry, String userName) {
         try {
+            User user = userService.findByUsername(userName);
             journalEntry.setDate(LocalDateTime.now());
-            journalEntryRepository.save(journalEntry);
+            JournalEntry saved = journalEntryRepository.save(journalEntry);
+
+            user.getJournalEntries().add(saved);
+            userService.saveEntry(user);
             System.out.println("Saved to MongoDB");
         }catch (Exception e){
-            log.error("Exception", e);
+            log.error(e.getMessage());
+            throw new RuntimeException("Error saving JournalEntry");
         }
+    }
 
+    public void saveEntry(JournalEntry journalEntry) {
+        journalEntryRepository.save(journalEntry);
     }
 
     public List<JournalEntry> findAll() {
@@ -36,8 +50,17 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id).get();
     }
 
-    public void deleteById(ObjectId id) {
-         journalEntryRepository.deleteById(id);
+    public void deleteById(ObjectId id, String name) {
+        User user = userService.findByUsername(name);
+
+        if (user == null) {
+            log.error("User not found with username: {}", name);
+            throw new RuntimeException("User not found");
+        }
+
+        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+        userService.saveEntry(user);
+        journalEntryRepository.deleteById(id);
     }
 
     public JournalEntry updateById(JournalEntry journalEntry) {
